requirements.txt
django==2.2
djangorestframework==3.12.1
requests==2.24.0
django-crispy-forms==1.9.2
pylint==2.6.0
flake8==3.8.4
yapf==0.30.0